Rename 'settings.py' to 'constants.py'
