+++ scholar/constants.py
